add first working version for evaluation
