+++ package.json
@@ -17,15 +17,17 @@
     "gatsby-plugin-sharp": "^2.0.15",
     "gatsby-remark-copy-linked-files": "^2.0.7",
     "gatsby-remark-images": "^3.0.1",
+    "gatsby-remark-prismjs": "^3.2.2",
     "gatsby-remark-relative-images": "^0.2.0",
     "gatsby-source-filesystem": "^2.0.12",
-    "gatsby-transformer-remark": "^2.1.15",
+    "gatsby-transformer-remark": "^2.2.1",
     "gatsby-transformer-sharp": "^2.1.9",
     "lodash": "^4.17.5",
     "lodash-webpack-plugin": "^0.11.4",
     "netlify-cms": "^2.3.1",
     "node-sass": "^4.11.0",
     "parcel-bundler": "^1.9.4",
+    "prismjs": "^1.15.0",
     "prop-types": "^15.6.0",
     "react": "^16.2.0",
     "react-dom": "^16.4.1",
@@ -45,7 +47,7 @@
     "serve": "gatsby serve",
     "format": "prettier --trailing-comma es5 --no-semi --single-quote --write \"{gatsby-*.js,src/**/*.js}\"",
     "test": "echo \"Error: no test specified\" && exit 1",
-    "scrap": "node -e 'require(\"./gatsby-node.js\").scrapFundamentals()'"
+    "scrap": "node -e 'require(\"./scrap-fundamentals.js\").scrapFundamentals()'"
   },
   "devDependencies": {
     "axios": "^0.18.0",
