+++ src/components/layout.scss
@@ -30,3 +30,76 @@ body {
   background-color: white;
   margin-left: calc((100% - 1440px)/2);
 }
+
+pre {
+    margin: 25px 0px;
+    // background-color: #fff !important;
+    white-space: pre-wrap;
+    white-space: -moz-pre-wrap;
+    white-space: -pre-wrap;
+    white-space: -o-pre-wrap;
+    word-wrap: break-word;
+    font-family: monospace;
+    display: block;
+    padding: 9.5px;
+    font-size: 13px;
+    line-height: 1.42857143;
+    color: #333;
+    word-break: break-all;
+    border: 1px solid #ccc;
+    border-radius: 4px;
+}
+
+.docs-component {
+    margin-bottom: -28px;
+    border-radius: 4px;
+    border-bottom-left-radius: 0;
+    border-bottom-right-radius: 0;
+    padding-bottom: 8px;
+    border: 1px solid #ccc;
+    position: relative;
+    background-color: #f3f4f5;
+    transition: background-color .25s;
+}
+
+.fd-tile {
+    background-color: #F3F4F5 !important;
+    font-size: 14px;
+    line-height: 1.42857;
+    font-family: '72', -apple-system, BlinkMacSystemFont, Segoe UI, Roboto, Helvetica Neue, Arial, sans-serif, Apple Color Emoji, Segoe UI Emoji, Segoe UI Symbol;
+    display: flex;
+    color: var(--fd-tile-color);
+}
+
+.docs-component__bg-toggle {
+    position: absolute;
+    right: 10px;
+    top: 10px;
+}
+
+.fd-form__label, .fd-form__legend {
+    font-family: '72', -apple-system, BlinkMacSystemFont, Segoe UI, Roboto, Helvetica Neue, Arial, sans-serif, Apple Color Emoji, Segoe UI Emoji, Segoe UI Symbol;
+    font-size: 1rem;
+    line-height: 1.42857;
+    font-weight: 400;
+    display: block;
+    margin-bottom: 8px;
+    border: 0;
+    color: #6a6d70;
+}
+
+.docs-component > .fd-tile__content {
+    padding: 60px 20px;
+    overflow: visible;
+}
+
+.fd-tile__content {
+    flex: 1;
+    align-self: center;
+}
+
+.token.tag {
+    font-size: 13px;
+    background-color: #fff;
+    font-family: monospace;
+}
